+++ src/ch.d.ts
@@ -11,6 +11,8 @@ type PasswordOptions = {
 export type { Comparator, PasswordOptions };
 
 declare function isObject<T = any>(o: any, empty?: boolean): o is object & T;
+declare function isProperty<T>(val: any, obj: T | T[]): boolean;
+
 declare function isNil(n: any): n is null | undefined;
 
 declare function isBoolean(b: any): b is boolean;
@@ -21,6 +23,7 @@ declare function isSymbol(s: any): s is symbol;
 declare function isFunction(func: any): func is Function;
 
 declare function isArray<T = any>(a: any, comp?: Comparator | null, len?: number | null): a is Array<T>;
+declare function isIn<T>(val: any, arr: T | T[]): boolean;
 
 declare function isAscii(c: any, ext?: boolean): c is number;
 declare function isInteger(n: any, type?: boolean): n is number;
@@ -59,8 +62,6 @@ declare function isValidDate(d: any, min?: Date, max?: Date): d is Date;
 declare function isTimestamp(t: any, type?: boolean): t is number;
 declare function isValidTimestamp(t: any, min?: number, max?: number, type?: boolean): t is number;
 
-declare function isCustomType<T>(val: any, customType: T | T[]): boolean;
-
 declare function ucfirst(s: string, everyWords?: boolean): string | false;
 declare function normalizeNickname(nickname: string, firstName: string, lastName: string): string | false;
 declare function normalizeName(s: string): string | false;
@@ -72,7 +73,9 @@ export { isBoolean,
          isJson,
          isFunction,
          isObject,
+         isProperty,
          isArray,
+         isIn,
          isAscii,
          isInteger,
          isValidInteger,
@@ -107,7 +110,6 @@ export { isBoolean,
          isTimestamp,
          isValidTimestamp,
          ucfirst,
-         isCustomType,
          normalizeNickname,
          normalizeName,
          normalizeEmail };