+++ src/structural.ts
@@ -4,7 +4,7 @@ function isObject<T = any>(o: any, empty = false): o is object & T {
   return o !== null && typeof o === "object" && !isArray(o) && (empty ? !!Object.keys(o).length : true);
 }
 
-function hasProperty<T>(val: any, obj: T): boolean {
+function isProperty<T>(val: any, obj: object): boolean {
   return isObject<T>(obj) ? Object.keys(obj).includes(val) : false;
 }
 
@@ -15,5 +15,5 @@ function isNil(n: any): n is null | undefined {
 export {
   isNil,
   isObject,
-  hasProperty,
+  isProperty,
 };